Create Connect-Printer.Ps1
This script connects a computer to a specified printer based on the closest print server. It first determines the primary site based on the closest domain controller and then selects the appropriate print server based on the site name. If the print server supports the Follow-You Print service, it automatically connects to that printer. Otherwise, it prompts the user to select a printer from a list.