Create Reset-CitrixPorifle.Ps1
This PowerShell script automates the process of resetting Citrix profiles, simplifying the task and enhancing overall efficiency. It intelligently detects the user profile server location and active sessions associated with the specified user. The script provides an option to request approval from the customer before proceeding with the profile reset. Once approval is obtained, the script terminates the active session if necessary and proceeds to reset the Citrix profile for the specified user.