Create Convert-PIMActiveToeligibleAssignments.Ps1
The Convert-PimEligibleAssignments function retrieves the eligible assignments for Azure Privileged Identity Management (PIM) roles. It uses a predefined group list to identify the groups for which the eligible assignments need to be retrieved. The function processes each group in the list, retrieves the Azure AD group corresponding to the group name, and then retrieves the eligible assignments for that group using the Get-AzureADMSPrivilegedRoleAssignment cmdlet. The eligible assignments are stored in an array of objects, and the results are exported to a CSV file and displayed on the console.